updated delegation tests and added delegation docs
