+++ Underscore.cs/Collection/Implementation/Delegation.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 using Underscore.Object.Reflection;
 
 namespace Underscore.Collection
@@ -16,6 +12,11 @@ namespace Underscore.Collection
             _methodComponent = methodComponent;
         }
 
+		/// <summary>
+		/// returns an IEnumerable consisting of the results of 
+		/// each item after having the method with
+		/// name methodName being called
+		/// </summary>
         public IEnumerable<T> Invoke<T>(IEnumerable<T> items, string methodName)
         {
             foreach (var item in items)
@@ -25,11 +26,16 @@ namespace Underscore.Collection
             }
         }
 
+		/// <summary>
+		/// returns an IEnumerable consisting of the results of 
+		/// each item after having the method with
+		/// name methodName being called with arguments
+		/// </summary>
         public IEnumerable<T> Invoke<T>(IEnumerable<T> items, string methodName, params object[] arguments)
         {
             foreach (var item in items)
             {
-                _methodComponent.Invoke(item, methodName,arguments);
+                _methodComponent.Invoke(item, methodName, arguments);
                 yield return item;
             }
         }
