+++ docs/api/Collection.md
@@ -26,7 +26,45 @@ _.Collection.Cycle(new int[] {1, 2, 3}) // returns {1, 2, 3, 1, 2, 3, 1, 2, 3, 1
 ```
 
 ## Delegation
-### TODO
+### IEnumerable\<T\> Invoke\<T\>(IEnumerable\<T\> items, string methodName)
+Returns an IEnumerable consisting of the results of each item after having the method with name methodName being called.
+```
+class Foo
+{
+    private static int counter = 0;
+    public int Value;
+
+    public void Bar() {
+        value = counter;
+        counter++;
+    }
+}
+
+Foo[] items = new Foo[] {new Foo(), new Foo(), new Foo(), new Foo()}
+
+IEnumerable<Foo> afterInvoke = _.Collection.Invoke(items, "Bar");
+
+afterInvoke.Select(foo => foo.Value); // returns {1, 2, 3, 4}
+```
+
+### IEnumerable\<T\> Invoke\<T\>(IEnumerable\<T\> items, string methodName, params object[] arguments)
+Returns an IEnumerable consisting of the results of each item after having the method with name methodName being called with the given arguments.
+```
+class Foo
+{
+    public int Value;
+
+    public void Bar(int n) {
+        value = n;
+    }
+}
+
+Foo[] items = new Foo[] {new Foo(), new Foo(), new Foo(), new Foo()}
+
+IEnumerable<Foo> afterInvoke = _.Collection.Invoke(items, "Bar", new object[] {1});
+
+afterInvoke.Select(foo => foo.Value); // returns {1, 1, 1, 1}
+```
 
 ## Partition
 ### IEnumerable\<IEnumerable\<T\>\> Chunk\<T\>(IEnumerable\<T\> collection, int size)
